+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * Make Wood Fence Gates Flammable
 * Make NonFlammable woodtypes unable to be used as fuel. Fixes #1249
 * Make Palo Verde Logs able to be used as fuel. Fixes #1250
+* Make Prairie Grass grow when bone mealing grass blocks in the Prairie Biomes. Closes #1044
 
 # 4.0.0.2
 * Fix broken NBT files. Fixes #1206 and #1207